backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv==1.0.0",
    "fastapi==0.124.0",
    "openai-agents==0.6.2",
    "qdrant-client>=1.16.1",
    "uvicorn==0.24.0",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "requests==2.31.0",
    "python-multipart==0.0.9",
    "pymongo>=4.15.5",
    "google-generativeai>=0.8.5",
]
